Gros commit :)
+ Redimensionnement des brains (adaptation des valeurs en pourcentage de la taille de la pluggrid)
   * Adaptation des widgets de rendu
   /!\ retrocompatibilité au niveau génome non implémentée

+ Suppression locale des sensors et effectors
    + possibilité des créer ou non un motorsEffector pour les axes de libertés

* Refactoring des Modifiers en Effectors (pas terminé)

* Interface (Images, mise en forme, etc.)

git-svn-id: http://www.genecraft-project.com/svn/genecraft/trunk@127 2ce2bb41-2ff2-4fd5-96a4-e3c75a7bd90e

+++ src/genelab/core/entities/brain/synapse.cpp
@@ -13,12 +13,11 @@ namespace GeneLabCore {
         QVariantList l = data.toList();
 
         foreach(QVariant v, l) {
-            int x = v.toMap()["x"].toInt();
-            int y = v.toMap()["y"].toInt();
+            float x = v.toMap()["x"].toDouble();
+            float y = v.toMap()["y"].toDouble();
             float w = v.toMap()["w"].toDouble();
             this->connectTo(x, y, w);
         }
-
     }
 
     QVariant Synapse::serialize() {
@@ -27,13 +26,12 @@ namespace GeneLabCore {
         for(int i = 0; i < this->connexions.size(); i++) {
             NeuralConnexion c = connexions[i];
             QVariantMap dataC;
-            dataC.insert("x", c.x);
-            dataC.insert("y", c.y);
+            dataC.insert("x", (double)c.x);
+            dataC.insert("y", (double)c.y);
             dataC.insert("w", (double)c.weight);
             l.append(dataC);
         }
 
         return l;
     }
-
 }
